requirements.txt
langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pytest>=7.0.0
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.8.0
tavily-python>=0.3.0
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
